pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-mask"
version = "0.1.0"
description = "Distributed code replay system for Upwork privacy"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Loomworks"}
]
keywords = ["automation", "code-replay", "upwork", "privacy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: X11 Applications",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Hardware",
]

dependencies = [
    "unidiff>=0.7.5",
    "pyyaml>=6.0",
    "python-xlib>=0.33",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
    "black>=23.0",
    "isort>=5.12",
    "flake8>=6.0",
]

[project.scripts]
mask-capture = "capture.cli:main"
mask-daemon = "orchestrator.session_orchestrator:main"

[project.urls]
Homepage = "https://github.com/loomworks/project-mask"
Documentation = "https://github.com/loomworks/project-mask#readme"
Repository = "https://github.com/loomworks/project-mask"

[tool.setuptools]
packages = ["capture", "replay", "upwork", "orchestrator", "config"]

[tool.setuptools.package-data]
config = ["*.yaml"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["capture", "replay", "upwork", "orchestrator", "config"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "unidiff.*",
    "Xlib.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
